backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
httpx==0.27.2
PyJWT[crypto]==2.9.0
